+++ app/src/main/java/com/example/recipe_provider/dto/IngredientEntity.java
@@ -1,17 +1,17 @@
 package com.example.recipe_provider.dto;
 
 public class IngredientEntity {
-    private int id;
+    private long id;
     private String name;
     private int remain;
 
-    public IngredientEntity(int id, String name, int remain) {
+    public IngredientEntity(long id, String name, int remain) {
         this.id = id;
         this.name = name;
         this.remain = remain;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
